Refactor: Split Components for Improved Maintainability and Readability
